- client: add a new mechanism for assigning coproc instances to tasks,         and passing them the corresponding --device N cmdline args.         This fixes a bug introduced in 17402 (Feb 26)         that broke the --device feature,         presumably causing problems on systems with multiple GPUs.
    client/
        app_start.cpp
        client_types.h
    lib/
        coproc.h

svn path=/branches/boinc_core_release_6_6a/; revision=17571

+++ client/client_types.h
@@ -412,7 +412,7 @@ struct APP_VERSION {
     char api_version[16];
     double avg_ncpus;
     double max_ncpus;
-    double ncudas;
+    int ncudas;
     double flops;
         /// additional cmdline args
     char cmdline[256];
